refactor extras into browsing package
